refactor(match): reorganize namespaces and remove unused folder for match classes
