+++ lib/DartsScorer.Main/DartsScorer.Main.csproj
@@ -6,8 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Folder Include="Models/Variants/" />
-  </ItemGroup>
-
 </Project>
