+++ lib/DartsScorer.Main/Match/MatchType.cs
@@ -1,4 +1,4 @@
-namespace DartsScorer.Main.Models.Variants;
+namespace DartsScorer.Main.Match;
 
 public enum MatchType
 {
