+++ lib/DartsScorer.Main/Match/x01/Match.cs
@@ -1,5 +1,5 @@
 
-namespace DartsScorer.Main.Models.Variants;
+namespace DartsScorer.Main.Match.x01;
 
 public class Match : CommonMatch
 {
