+++ lib/tests/DartsScorer.Tests/MatchTests.cs
@@ -1,5 +1,4 @@
-using System;
-using DartsScorer.Main.Models.Variants;
+using DartsScorer.Main;
 
 namespace DartsScorer.Tests;
 
@@ -8,26 +7,26 @@ public class MatchTests
     [Test]
     public void Match_Instantiate_Match_Success()
     {
-        var match = new Match(Main.Models.Variants.MatchType.x01);
+        var match = new Match(Main.Match.MatchType.x01);
 
-        Assert.That(match.MatchType, Is.EqualTo(Main.Models.Variants.MatchType.x01));
+        Assert.That(match.MatchType, Is.EqualTo(Main.Match.MatchType.x01));
     }
 
     [Test]
     public void Match_Instantiate_Match_Fail()
     {
-        var match = new Match(Main.Models.Variants.MatchType.x01);
+        var match = new Match(Main.Match.MatchType.x01);
 
-        Assert.That(match.MatchType, Is.Not.EqualTo(Main.Models.Variants.MatchType.Killer));
+        Assert.That(match.MatchType, Is.Not.EqualTo(Main.Match.MatchType.Killer));
     }
 }
 
 internal class Match
 {
-    public Match(Main.Models.Variants.MatchType gameVariant)
+    public Match(Main.Match.MatchType matchType)
     {
-        MatchType = gameVariant;
+        MatchType = matchType;
     }
 
-    public Main.Models.Variants.MatchType MatchType { get; internal set; }
+    public Main.Match.MatchType MatchType { get; internal set; }
 }
